refactor: split kiosk handlers by responsibility
